+++ ui-test/smoke/resources.spec.ts
